+++ src/pydvl/value/__init__.py
@@ -1,0 +1,6 @@
+r"""
+Algorithms for the exact and approximate computation of value and semi-value.
+
+See :ref:`data valuation` for an introduction to the concepts and methods
+implemented here.
+"""