+++ src/pydvl/value/loo/naive.py
@@ -8,9 +8,7 @@ from pydvl.utils import Utility, maybe_progress
 __all__ = ["naive_loo"]
 
 
-def naive_loo(
-    u: Utility, *, progress: bool = True, **kwargs
-) -> OrderedDict[int, float]:
+def naive_loo(u: Utility, *, progress: bool = True) -> OrderedDict[int, float]:
     """Computes leave one out score.
 
     No caching nor parallelization is implemented.
@@ -20,8 +18,9 @@ def naive_loo(
     """
 
     values = {i: 0.0 for i in u.data.indices}
+    all_indices = set(u.data.indices)
     for i in maybe_progress(data.indices, progress):  # type: ignore
-        subset = np.setxor1d(u.data.indices, [i], assume_unique=True)
+        subset = all_indices.difference({i})
         values[i] = u(u.data.indices) - u(subset)
 
     return sort_values(values)